Stack even lower
Instead of just tracking the maximum height of the board, we now
consider whether e.g. a J tetromino is standing up-right or is laying
on its side. Laying on its side will have a higher score, because more
of its tils are lower overall.

This bot reaches over 300 lines.